+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Added `BLEAdvertisedDevice::adv_type`, `BLEAdvertisedDevice::adv_flags`
 - Added `BLEAddress::from_str`
 - Added whitelist API.
+- Added `BLEClient::get_rssi` ([#58](https::/github.com/taks/esp32-nimble/pull/58))
+- Added `BLEConnDesc`
+- Fixed no_std build.
 
 ## [0.4.0] - 2023-12-01
 - Added `BLEAdvertising::min_interval`, `BLEAdvertising::max_interval` ([#51](https::/github.com/taks/esp32-nimble/pull/51))