+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.4.1]
+## [0.5.0] - 2024-01-10
 - Added `BLEScan::find_device` ([#55](https::/github.com/taks/esp32-nimble/pull/55))
 - Added `BLEAdvertisedDevice::adv_type`, `BLEAdvertisedDevice::adv_flags`
 - Added `BLEAddress::from_str`